fix: nft detail owner
